Running an exec command now returns the response to that command.
My server on connection would show the text "Waiting for Command". In telnet I would then type "auth user pass". Then the server would return "Authorized".

Using this class I would connect to my server, and have the following code:

    $response = $telnet->exec('auth user pass');

And $response would be equal to "Waiting for Command". This is because the server I am telneting into uses \n before prompting instead of $. 

So I added an initial waitPrompt() call to ensure that once I started to execute commands I would get the response to that command and not to the previous one.